Changed security.

+++ src/main/java/com/tms/lingofriends/service/CourseService.java
@@ -1,25 +1,29 @@
 package com.tms.lingofriends.service;
 
+import com.tms.lingofriends.exception.AccessException;
 import com.tms.lingofriends.exception.NotFoundException;
 import com.tms.lingofriends.mapper.CourseToCourseResponseMapper;
 import com.tms.lingofriends.model.Course;
 import com.tms.lingofriends.model.response.CourseResponse;
 import com.tms.lingofriends.repository.CourseRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.sql.Timestamp;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import static com.tms.lingofriends.util.ExceptionMesseges.ACCESS_IS_DENIED;
 import static com.tms.lingofriends.util.ExceptionMesseges.COURSES_NOT_FOUND;
 import static com.tms.lingofriends.util.ExceptionMesseges.COURSE_NOT_FOUND;
 
 @Service
 public class CourseService {
-    CourseRepository courseRepository;
-    CourseToCourseResponseMapper courseToCourseResponseMapper;
+    private final CourseRepository courseRepository;
+    private final CourseToCourseResponseMapper courseToCourseResponseMapper;
 
     @Autowired
     public CourseService(CourseRepository courseRepository, CourseToCourseResponseMapper courseToCourseResponseMapper) {
@@ -36,7 +40,6 @@ public class CourseService {
         }
     }
 
-    // for user get all course
     public List<CourseResponse> getAllCoursesResponse() throws NotFoundException {
         List<CourseResponse> courses = courseRepository.findAll().stream()
                 .filter(course -> !course.isDeleted())
@@ -47,13 +50,11 @@ public class CourseService {
         } else throw new NotFoundException(COURSES_NOT_FOUND);
     }
 
-    // for admin course by id
     public Course getCourseById(int id) {
         return courseRepository.findById(id).orElseThrow(() ->
                 new NotFoundException(COURSE_NOT_FOUND));
     }
 
-    // for user course by id
     public CourseResponse getCourseResponseById(int id) {
         Optional<Course> course = courseRepository.findById(id);
         if (!course.isEmpty()) {
@@ -73,7 +74,6 @@ public class CourseService {
         }
     }
 
-    // for admin by user id
     public List<Course> findCourseByUserId(Integer userId) {
         List<Course> courses = courseRepository.findCourseByUserId(userId);
         if (!courses.isEmpty()) {
@@ -96,15 +96,35 @@ public class CourseService {
     }
 
     public Course createCourse(Course course) {
+        course.setCreated(new Timestamp(System.currentTimeMillis()));
+        course.setChanged(new Timestamp(System.currentTimeMillis()));
         return courseRepository.save(course);
     }
 
     public Course updateCourse(Course course) {
-        return courseRepository.saveAndFlush(course);
+        Course courses = courseRepository.findById(course.getId()).get();
+        if (authorization(courses.getUserLogin())) {
+            return courseRepository.saveAndFlush(course);
+        } else {
+            throw new AccessException(ACCESS_IS_DENIED);
+        }
     }
 
     @Transactional
     public void deleteCourseById(int id) {
-        courseRepository.deleteCourseById(id);
+        if (authorization(getUserLogin(id))) {
+            courseRepository.deleteCourseById(id);
+        } else {
+            throw new AccessException(ACCESS_IS_DENIED);
+        }
+    }
+
+    private String getUserLogin(int id) {
+        return courseRepository.findById(id).get().getUserLogin();
+    }
+
+    public boolean authorization(String login) {
+        return SecurityContextHolder.getContext()
+                .getAuthentication().getName().equals(login);
     }
 }
