+++ src/main/java/com/tms/lingofriends/service/LessonService.java
@@ -1,5 +1,6 @@
 package com.tms.lingofriends.service;
 
+import com.tms.lingofriends.exception.AccessException;
 import com.tms.lingofriends.exception.NotFoundException;
 import com.tms.lingofriends.mapper.LessonToLessonResponseMapper;
 import com.tms.lingofriends.model.Lesson;
@@ -7,19 +8,22 @@ import com.tms.lingofriends.model.response.LessonResponse;
 import com.tms.lingofriends.repository.LessonRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import java.sql.Timestamp;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import static com.tms.lingofriends.util.ExceptionMesseges.ACCESS_IS_DENIED;
 import static com.tms.lingofriends.util.ExceptionMesseges.LESSON_NOT_FOUND;
 import static com.tms.lingofriends.util.ExceptionMesseges.LESSONS_NOT_FOUND;
 
 @Service
 public class LessonService {
-    LessonRepository lessonRepository;
-    LessonToLessonResponseMapper lessonToLessonResponseMapper;
+    private final LessonRepository lessonRepository;
+    private final LessonToLessonResponseMapper lessonToLessonResponseMapper;
 
     @Autowired
     public LessonService(LessonRepository lessonRepository, LessonToLessonResponseMapper lessonToLessonResponseMapper) {
@@ -27,7 +31,6 @@ public class LessonService {
         this.lessonToLessonResponseMapper = lessonToLessonResponseMapper;
     }
 
-    // for admin get all lesson
     public List<Lesson> getAllLesson() {
         List<Lesson> lessons = lessonRepository.findAll();
         if (!lessons.isEmpty()) {
@@ -37,7 +40,6 @@ public class LessonService {
         }
     }
 
-    // for user get all lesson
     public List<LessonResponse> getAllLessonResponse() throws NotFoundException {
         List<LessonResponse> lessons = lessonRepository.findAll().stream()
                 .filter(lesson -> !lesson.isDeleted())
@@ -48,13 +50,11 @@ public class LessonService {
         } else throw new NotFoundException(LESSONS_NOT_FOUND);
     }
 
-    // for admin get by id
     public Lesson getLessonById(int id) {
         return lessonRepository.findById(id).orElseThrow(() ->
                 new NotFoundException(LESSON_NOT_FOUND));
     }
 
-    // for user get by id
     public LessonResponse getLessonResponseById(int id) {
         Optional<Lesson> lesson = lessonRepository.findById(id);
         if (!lesson.isEmpty()) {
@@ -71,7 +71,6 @@ public class LessonService {
         }
     }
 
-    // for user by course id
     public List<LessonResponse> findLessonResponseByCourseId(int courseId) {
         List<LessonResponse> lessons = lessonRepository.findLessonByCourseId(courseId).stream()
                 .filter(lesson -> !lesson.isDeleted())
@@ -97,16 +96,35 @@ public class LessonService {
     }
 
     public Lesson createLesson(Lesson lesson) {
+        lesson.setCreated(new Timestamp(System.currentTimeMillis()));
+        lesson.setChanged(new Timestamp(System.currentTimeMillis()));
         return lessonRepository.save(lesson);
     }
 
     public Lesson updateLesson(Lesson lesson) {
-        return lessonRepository.saveAndFlush(lesson);
+        Lesson lessons = lessonRepository.findById(lesson.getId()).get();
+        if (authorization(lesson.getUserLogin())) {
+            return lessonRepository.saveAndFlush(lesson);
+        } else {
+            throw new AccessException(ACCESS_IS_DENIED);
+        }
     }
 
     @Transactional
     public void deleteLessonById(int id) {
-        lessonRepository.deleteLessonById(id);
+        if (authorization(getUserLogin(id))) {
+            lessonRepository.deleteLessonById(id);
+        } else {
+            throw new AccessException(ACCESS_IS_DENIED);
+        }
     }
 
+    private String getUserLogin(int id) {
+        return lessonRepository.findById(id).get().getUserLogin();
+    }
+
+    public boolean authorization(String login) {
+        return SecurityContextHolder.getContext()
+                .getAuthentication().getName().equals(login);
+    }
 }
