+++ src/main/java/com/tms/lingofriends/service/SubscriptionService.java
@@ -1,5 +1,6 @@
 package com.tms.lingofriends.service;
 
+import com.tms.lingofriends.exception.AccessException;
 import com.tms.lingofriends.exception.NotFoundException;
 import com.tms.lingofriends.mapper.SubscriptionToSubscriptionResponseMapper;
 import com.tms.lingofriends.model.Subscription;
@@ -7,6 +8,7 @@ import com.tms.lingofriends.model.response.SubscriptionResponse;
 import com.tms.lingofriends.repository.SubscriptionRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
@@ -14,20 +16,22 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import static com.tms.lingofriends.util.ExceptionMesseges.ACCESS_IS_DENIED;
 import static com.tms.lingofriends.util.ExceptionMesseges.SUBSCRIPTION_NOT_FOUND;
 import static com.tms.lingofriends.util.ExceptionMesseges.SUBSCRIPTIONS_NOT_FOUND;
 
 
 @Service
 public class SubscriptionService {
-    SubscriptionRepository subscriptionRepository;
-    SubscriptionToSubscriptionResponseMapper subscriptionToSubscriptionResponseMapper;
+    private final SubscriptionRepository subscriptionRepository;
+    private final SubscriptionToSubscriptionResponseMapper subscriptionToSubscriptionResponseMapper;
 
     @Autowired
-    public SubscriptionService(SubscriptionRepository subscriptionRepository,SubscriptionToSubscriptionResponseMapper subscriptionToSubscriptionResponseMapper) {
+    public SubscriptionService(SubscriptionRepository subscriptionRepository, SubscriptionToSubscriptionResponseMapper subscriptionToSubscriptionResponseMapper) {
         this.subscriptionRepository = subscriptionRepository;
         this.subscriptionToSubscriptionResponseMapper = subscriptionToSubscriptionResponseMapper;
     }
+
     public List<Subscription> getAllSubscription() {
         List<Subscription> subscriptions = subscriptionRepository.findAll();
         if (!subscriptions.isEmpty()) {
@@ -36,16 +40,19 @@ public class SubscriptionService {
             throw new NotFoundException(SUBSCRIPTIONS_NOT_FOUND);
         }
     }
+
     public Subscription getSubscriptionById(int id) {
         return subscriptionRepository.findById(id).orElseThrow(() ->
                 new NotFoundException(SUBSCRIPTION_NOT_FOUND));
     }
+
     public SubscriptionResponse getSubscriptionResponseById(int id) {
         Optional<Subscription> subscription = subscriptionRepository.findById(id);
         if (!subscription.isEmpty()) {
             return subscriptionToSubscriptionResponseMapper.subscriptionToResponse(subscription.get());
         } else throw new NotFoundException(SUBSCRIPTION_NOT_FOUND);
     }
+
     public List<SubscriptionResponse> findSubscriptionResponseByUserId(int userId) {
         List<SubscriptionResponse> subscriptions = subscriptionRepository.findSubscriptionByUserId(userId).stream()
                 .map(subscriptionToSubscriptionResponseMapper::subscriptionToResponse)
@@ -56,6 +63,7 @@ public class SubscriptionService {
             throw new NotFoundException(SUBSCRIPTION_NOT_FOUND);
         }
     }
+
     public Subscription createSubscription(Subscription subscription) {
         subscription.setExpireDate(LocalDate.now().plusYears(1));
         subscription.setStatus(true);
@@ -63,11 +71,29 @@ public class SubscriptionService {
     }
 
     public Subscription updateCourse(Subscription subscription) {
-        return subscriptionRepository.saveAndFlush(subscription);
+        Subscription subscription1 = subscriptionRepository.findById(subscription.getId()).get();
+        if (authorization(subscription1.getUserLogin())) {
+            return subscriptionRepository.saveAndFlush(subscription);
+        } else {
+            throw new AccessException(ACCESS_IS_DENIED);
+        }
     }
 
     @Transactional
     public void deleteSubscriptionById(int id) {
-        subscriptionRepository.deleteSubscriptionById(id);
+        if (authorization(getUserLogin(id))) {
+            subscriptionRepository.deleteSubscriptionById(id);
+        } else {
+            throw new AccessException(ACCESS_IS_DENIED);
+        }
+    }
+
+    private String getUserLogin(int id) {
+        return subscriptionRepository.findById(id).get().getUserLogin();
+    }
+
+    public boolean authorization(String login) {
+        return SecurityContextHolder.getContext()
+                .getAuthentication().getName().equals(login);
     }
 }
